@@ -229,28 +373,43 @@ def awscrt_ext():
         sources=glob.glob('source/*.c'),
         extra_compile_args=extra_compile_args,
         extra_link_args=extra_link_args,
-        extra_objects=extra_objects
+        extra_objects=extra_objects,
+        define_macros=define_macros,
+        py_limited_api=py_limited_api,
     )
+
+
+def _load_readme():
+    readme_path = os.path.join(PROJECT_DIR, 'README.md')
+    with codecs.open(readme_path, 'r', 'utf-8') as f:
+        return f.read()
+
+
+def _load_version():
+    init_path = os.path.join(PROJECT_DIR, 'awscrt', '__init__.py')
+    with open(init_path) as fp:
+        return VERSION_RE.match(fp.read()).group(1)
 
 
 setuptools.setup(
     name="awscrt",
-    version="1.0.0-dev",
+    version=_load_version(),
+    license="Apache 2.0",
     author="Amazon Web Services, Inc",
     author_email="aws-sdk-common-runtime@amazon.com",
     description="A common runtime for AWS Python projects",
+    long_description=_load_readme(),
+    long_description_content_type='text/markdown',
     url="https://github.com/awslabs/aws-crt-python",
-    packages=['awscrt'],
+    # Note: find_packages() without extra args will end up installing test/
+    packages=setuptools.find_packages(include=['awscrt*']),
     classifiers=[
         "Programming Language :: Python :: 3",
         "License :: OSI Approved :: Apache Software License",
         "Operating System :: OS Independent",
     ],
-    python_requires='>=3.5',
+    python_requires='>=3.7',
     ext_modules=[awscrt_ext()],
-    cmdclass={'build_ext': awscrt_build_ext},
+    cmdclass={'build_ext': awscrt_build_ext, "bdist_wheel": bdist_wheel_abi3},
     test_suite='test',
-    tests_require=[
-        'boto3'
-    ]
 )
