@@ -8,12 +8,11 @@ Long-running event-loop threads are used for concurrency.
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0.
 
-from __future__ import absolute_import
 import _awscrt
 from awscrt import NativeResource
 from enum import IntEnum
-import io
 import threading
+from typing import Union
 
 
 class LogLevel(IntEnum):
@@ -48,8 +47,13 @@ class EventLoopGroup(NativeResource):
     need to do async work will ask the EventLoopGroup for an event-loop to use.
 
     Args:
-        num_threads (int): Number of event-loops to create.
-            Pass 0 to create one for each processor on the machine.
+        num_threads (Optional[int]): Maximum number of event-loops to create.
+            If unspecified, one is created for each processor on the machine.
+
+        cpu_group (Optional[int]): Optional processor group to which all
+            threads will be pinned. Useful for systems with non-uniform
+            memory access (NUMA) nodes. If specified, the number of threads
+            will be capped at the number of processors in the group.
 
     Attributes:
         shutdown_event (threading.Event): Signals when EventLoopGroup's threads
@@ -57,10 +61,22 @@ class EventLoopGroup(NativeResource):
             EventLoopGroup object is destroyed.
     """
 
+    _static_event_loop_group = None
+    _static_event_loop_group_lock = threading.Lock()
     __slots__ = ('shutdown_event')
 
-    def __init__(self, num_threads=0):
+    def __init__(self, num_threads=None, cpu_group=None):
         super().__init__()
+
+        if num_threads is None:
+            # C uses 0 to indicate defaults
+            num_threads = 0
+
+        if cpu_group is None:
+            is_pinned = False
+            cpu_group = 0
+        else:
+            is_pinned = True
 
         shutdown_event = threading.Event()
 
@@ -68,7 +84,19 @@ class EventLoopGroup(NativeResource):
             shutdown_event.set()
 
         self.shutdown_event = shutdown_event
-        self._binding = _awscrt.event_loop_group_new(num_threads, on_shutdown)
+        self._binding = _awscrt.event_loop_group_new(num_threads, is_pinned, cpu_group, on_shutdown)
+
+    @staticmethod
+    def get_or_create_static_default():
+        with EventLoopGroup._static_event_loop_group_lock:
+            if EventLoopGroup._static_event_loop_group is None:
+                EventLoopGroup._static_event_loop_group = EventLoopGroup()
+            return EventLoopGroup._static_event_loop_group
+
+    @staticmethod
+    def release_static_default():
+        with EventLoopGroup._static_event_loop_group_lock:
+            EventLoopGroup._static_event_loop_group = None
 
 
 class HostResolverBase(NativeResource):
@@ -83,6 +111,9 @@ class DefaultHostResolver(HostResolverBase):
         event_loop_group (EventLoopGroup): EventLoopGroup to use.
         max_hosts(int): Max host names to cache.
     """
+
+    _static_host_resolver = None
+    _static_host_resolver_lock = threading.Lock()
     __slots__ = ()
 
     def __init__(self, event_loop_group, max_hosts=16):
@@ -90,6 +121,19 @@ class DefaultHostResolver(HostResolverBase):
 
         super().__init__()
         self._binding = _awscrt.host_resolver_new_default(max_hosts, event_loop_group)
+
+    @staticmethod
+    def get_or_create_static_default():
+        with DefaultHostResolver._static_host_resolver_lock:
+            if DefaultHostResolver._static_host_resolver is None:
+                DefaultHostResolver._static_host_resolver = DefaultHostResolver(
+                    EventLoopGroup.get_or_create_static_default())
+            return DefaultHostResolver._static_host_resolver
+
+    @staticmethod
+    def release_static_default():
+        with DefaultHostResolver._static_host_resolver_lock:
+            DefaultHostResolver._static_host_resolver = None
 
 
 class ClientBootstrap(NativeResource):
@@ -104,6 +148,9 @@ class ClientBootstrap(NativeResource):
             internal resources finish shutting down.
             Shutdown begins when the ClientBootstrap object is destroyed.
     """
+
+    _static_client_bootstrap = None
+    _static_client_bootstrap_lock = threading.Lock()
     __slots__ = ('shutdown_event')
 
     def __init__(self, event_loop_group, host_resolver):
@@ -119,6 +166,20 @@ class ClientBootstrap(NativeResource):
 
         self.shutdown_event = shutdown_event
         self._binding = _awscrt.client_bootstrap_new(event_loop_group, host_resolver, on_shutdown)
+
+    @staticmethod
+    def get_or_create_static_default():
+        with ClientBootstrap._static_client_bootstrap_lock:
+            if ClientBootstrap._static_client_bootstrap is None:
+                ClientBootstrap._static_client_bootstrap = ClientBootstrap(
+                    EventLoopGroup.get_or_create_static_default(),
+                    DefaultHostResolver.get_or_create_static_default())
+            return ClientBootstrap._static_client_bootstrap
+
+    @staticmethod
+    def release_static_default():
+        with ClientBootstrap._static_client_bootstrap_lock:
+            ClientBootstrap._static_client_bootstrap = None
 
 
 def _read_binary_file(filepath):
@@ -190,6 +251,34 @@ class TlsVersion(IntEnum):
     DEFAULT = 128  #:
 
 
+class TlsCipherPref(IntEnum):
+    """TLS Cipher Preference.
+
+       Each TlsCipherPref represents an ordered list of TLS Ciphers to use when negotiating a TLS Connection. At
+       present, the ability to configure arbitrary orderings of TLS Ciphers is not allowed, and only a curated list of
+       vetted TlsCipherPref's are exposed."""
+
+    DEFAULT = 0
+    """The underlying platform's default TLS Cipher Preference ordering. This is usually the best option, as it will be
+       automatically updated as the underlying OS or platform changes, and will always be supported on all platforms."""
+
+    PQ_TLSv1_0_2021_05 = 6  #:
+    """A TLS Cipher Preference ordering that supports TLS 1.0 through TLS 1.3, and has Kyber Round 3 as its highest
+       priority post-quantum key exchange algorithm. PQ algorithms in this preference list will always be used in hybrid
+       mode, and will be combined with a classical ECDHE key exchange that is performed in addition to the PQ key
+       exchange. This preference makes a best-effort to negotiate a PQ algorithm, but if the peer does not support any
+       PQ algorithms the TLS connection will fall back to a single classical algorithm for key exchange (such as ECDHE
+       or RSA).
+
+       NIST has announced that they plan to eventually standardize Kyber. However, the NIST standardization process might
+       introduce minor changes that could cause the final Kyber standard to differ from the Kyber Round 3 implementation
+       available in this preference list."""
+
+    def is_supported(self):
+        """Return whether this Cipher Preference is available in the underlying platform's TLS implementation"""
+        return _awscrt.is_tls_cipher_supported(self.value)
+
+
 class TlsContextOptions:
     """Options to create a TLS context.
 
@@ -199,6 +288,7 @@ class TlsContextOptions:
     Attributes:
         min_tls_ver (TlsVersion): Minimum TLS version to use.
             System defaults are used by default.
+        cipher_pref (TlsCipherPref): The TLS Cipher Preference to use. System defaults are used by default.
         verify_peer (bool): Whether to validate the peer's x.509 certificate.
         alpn_list (Optional[List[str]]): If set, names to use in Application Layer
             Protocol Negotiation (ALPN). ALPN is not supported on all systems,
@@ -206,9 +296,25 @@ class TlsContextOptions:
             via :meth:`TlsConnectionOptions.set_alpn_list()`.
     """
     __slots__ = (
-        'min_tls_ver', 'ca_dirpath', 'ca_buffer', 'alpn_list',
-        'certificate_buffer', 'private_key_buffer',
-        'pkcs12_filepath', 'pkcs12_password', 'verify_peer')
+        'min_tls_ver',
+        'ca_dirpath',
+        'ca_buffer',
+        'cipher_pref',
+        'alpn_list',
+        'certificate_buffer',
+        'private_key_buffer',
+        'pkcs12_filepath',
+        'pkcs12_password',
+        'verify_peer',
+        '_pkcs11_lib',
+        '_pkcs11_user_pin',
+        '_pkcs11_slot_id',
+        '_pkcs11_token_label',
+        '_pkcs11_private_key_label',
+        '_pkcs11_cert_file_path',
+        '_pkcs11_cert_file_contents',
+        '_windows_cert_store_path',
+    )
 
     def __init__(self):
 
@@ -216,6 +322,7 @@ class TlsContextOptions:
             setattr(self, slot, None)
 
         self.min_tls_ver = TlsVersion.DEFAULT
+        self.cipher_pref = TlsCipherPref.DEFAULT
         self.verify_peer = True
 
     @staticmethod
@@ -263,7 +370,65 @@ class TlsContextOptions:
         opt.certificate_buffer = cert_buffer
         opt.private_key_buffer = key_buffer
 
-        opt.verify_peer = True
+        return opt
+
+    @staticmethod
+    def create_client_with_mtls_pkcs11(*,
+                                       pkcs11_lib: 'Pkcs11Lib',
+                                       user_pin: Union[str, None],
+                                       slot_id: int = None,
+                                       token_label: str = None,
+                                       private_key_label: str = None,
+                                       cert_file_path: str = None,
+                                       cert_file_contents=None):
+        """
+        Create options configured for use with mutual TLS in client mode,
+        using a PKCS#11 library for private key operations.
+
+        NOTE: This configuration only works on Unix devices.
+
+        Keyword Args:
+            pkcs11_lib (Pkcs11Lib): Use this PKCS#11 library
+
+            user_pin (str): User PIN, for logging into the PKCS#11 token.
+                Pass `None` to log into a token with a "protected authentication path".
+
+            slot_id (Optional[int]): ID of slot containing PKCS#11 token.
+                If not specified, the token will be chosen based on other criteria (such as token label).
+
+            token_label (Optional[str]): Label of the PKCS#11 token to use.
+                If not specified, the token will be chosen based on other criteria (such as slot ID).
+
+            private_key_label (Optional[str]): Label of private key object on PKCS#11 token.
+                If not specified, the key will be chosen based on other criteria
+                (such as being the only available private key on the token).
+
+            cert_file_path (Optional[str]): Use this X.509 certificate (path to file on disk).
+                The certificate must be PEM-formatted. The certificate may be
+                specified by other means instead (ex: `cert_file_contents`)
+
+            cert_file_contents (Optional[Union[str, bytes, bytearray]]):
+                Use this X.509 certificate (contents in memory).
+                The certificate must be PEM-formatted. The certificate may be
+                specified by other means instead (ex: `cert_file_path`)
+        """
+
+        assert isinstance(pkcs11_lib, Pkcs11Lib)
+        assert isinstance(user_pin, str) or user_pin is None
+        assert isinstance(slot_id, int) or slot_id is None
+        assert isinstance(token_label, str) or token_label is None
+        assert isinstance(private_key_label, str) or private_key_label is None
+        assert isinstance(cert_file_path, str) or cert_file_path is None
+        # note: not validating cert_file_contents, because "bytes-like object" isn't a strict type
+
+        opt = TlsContextOptions()
+        opt._pkcs11_lib = pkcs11_lib
+        opt._pkcs11_user_pin = user_pin
+        opt._pkcs11_slot_id = slot_id
+        opt._pkcs11_token_label = token_label
+        opt._pkcs11_private_key_label = private_key_label
+        opt._pkcs11_cert_file_path = cert_file_path
+        opt._pkcs11_cert_file_contents = cert_file_contents
         return opt
 
     @staticmethod
@@ -288,7 +453,27 @@ class TlsContextOptions:
         opt = TlsContextOptions()
         opt.pkcs12_filepath = pkcs12_filepath
         opt.pkcs12_password = pkcs12_password
-        opt.verify_peer = True
+        return opt
+
+    @staticmethod
+    def create_client_with_mtls_windows_cert_store_path(cert_path):
+        """
+        Create options configured for use with mutual TLS in client mode,
+        using a certificate in a Windows certificate store.
+
+        NOTE: This configuration only works on Windows devices.
+
+        Args:
+            cert_path (str): Path to certificate in a Windows certificate store.
+                The path must use backslashes and end with the certificate's thumbprint.
+                Example: ``CurrentUser\\MY\\A11F8A9B5DF5B98BA3508FBCA575D09570E0D2C6``
+
+        Returns:
+            TlsContextOptions
+        """
+        assert isinstance(cert_path, str)
+        opt = TlsContextOptions()
+        opt._windows_cert_store_path = cert_path
         return opt
 
     @staticmethod
@@ -410,6 +595,7 @@ class ClientTlsContext(NativeResource):
         super().__init__()
         self._binding = _awscrt.client_tls_ctx_new(
             options.min_tls_ver.value,
+            options.cipher_pref.value,
             options.ca_dirpath,
             options.ca_buffer,
             _alpn_list_to_str(options.alpn_list),
@@ -417,7 +603,15 @@ class ClientTlsContext(NativeResource):
             options.private_key_buffer,
             options.pkcs12_filepath,
             options.pkcs12_password,
-            options.verify_peer
+            options.verify_peer,
+            options._pkcs11_lib,
+            options._pkcs11_user_pin,
+            options._pkcs11_slot_id,
+            options._pkcs11_token_label,
+            options._pkcs11_private_key_label,
+            options._pkcs11_cert_file_path,
+            options._pkcs11_cert_file_contents,
+            options._windows_cert_store_path,
         )
 
     def new_connection_options(self):
@@ -490,20 +684,47 @@ def is_alpn_available():
 
 
 class InputStream(NativeResource):
-    """InputStream allows `awscrt` native code to read from Python I/O classes.
+    """InputStream allows `awscrt` native code to read from Python binary I/O classes.
 
     Args:
-        stream (io.IOBase): Python I/O stream to wrap.
+        stream (io.IOBase): Python binary I/O stream to wrap.
     """
-    __slots__ = ()
+    __slots__ = ('_stream')
     # TODO: Implement IOBase interface so Python can read from this class as well.
 
     def __init__(self, stream):
-        assert isinstance(stream, io.IOBase)
+        # duck-type instead of checking inheritance from IOBase.
+        # At the least, stream must have read()
+        if not callable(getattr(stream, 'read', None)):
+            raise TypeError('I/O stream type expected')
         assert not isinstance(stream, InputStream)
 
         super().__init__()
-        self._binding = _awscrt.input_stream_new(stream)
+        self._stream = stream
+        self._binding = _awscrt.input_stream_new(self)
+
+    def _read_into_memoryview(self, m):
+        # Read into memoryview m.
+        # Return number of bytes read, or None if no data available.
+        try:
+            # prefer the most efficient read methods,
+            if hasattr(self._stream, 'readinto1'):
+                return self._stream.readinto1(m)
+            if hasattr(self._stream, 'readinto'):
+                return self._stream.readinto(m)
+
+            if hasattr(self._stream, 'read1'):
+                data = self._stream.read1(len(m))
+            else:
+                data = self._stream.read(len(m))
+            n = len(data)
+            m[:n] = data
+            return n
+        except BlockingIOError:
+            return None
+
+    def _seek(self, offset, whence):
+        return self._stream.seek(offset, whence)
 
     @classmethod
     def wrap(cls, stream, allow_none=False):
@@ -511,7 +732,7 @@ class InputStream(NativeResource):
         Given some stream type, returns an :class:`InputStream`.
 
         Args:
-            stream (Union[io.IOBase, InputStream, None]): I/O stream to wrap.
+            stream (Union[io.IOBase, InputStream, None]): Binary I/O stream to wrap.
             allow_none (bool): Whether to allow `stream` to be None.
                 If False (default), and `stream` is None, an exception is raised.
 
@@ -520,10 +741,64 @@ class InputStream(NativeResource):
             Otherwise, an :class:`InputStream` which wraps the `stream` is returned.
             If `allow_none` is True, and `stream` is None, then None is returned.
         """
-        if isinstance(stream, InputStream):
-            return stream
-        if isinstance(stream, io.IOBase):
-            return cls(stream)
         if stream is None and allow_none:
             return None
-        raise TypeError('I/O stream type expected')
+        if isinstance(stream, InputStream):
+            return stream
+        return cls(stream)
+
+
+class Pkcs11Lib(NativeResource):
+    """
+    Handle to a loaded PKCS#11 library.
+
+    For most use cases, a single instance of :class:`Pkcs11Lib` should be used for the
+    lifetime of your application.
+
+    Keyword Args:
+        file (str): Path to PKCS#11 library.
+        behavior (Optional[InitializeFinalizeBehavior]):
+            Specifies how `C_Initialize()` and `C_Finalize()` will be called
+            on the PKCS#11 library (default is :attr:`InitializeFinalizeBehavior.DEFAULT`)
+    """
+
+    class InitializeFinalizeBehavior(IntEnum):
+        """
+        An enumeration.
+
+        Controls how `C_Initialize()` and `C_Finalize()` are called on the PKCS#11 library.
+        """
+
+        DEFAULT = 0
+        """
+        Relaxed behavior that accommodates most use cases.
+
+        `C_Initialize()` is called on creation, and "already-initialized"
+        errors are ignored. `C_Finalize()` is never called, just in case
+        another part of your application is still using the PKCS#11 library.
+        """
+
+        OMIT = 1
+        """
+        Skip calling `C_Initialize()` and `C_Finalize()`.
+
+        Use this if your application has already initialized the PKCS#11 library, and
+        you do not want `C_Initialize()` called again.
+        """
+
+        STRICT = 2
+        """
+        `C_Initialize()` is called on creation and `C_Finalize()` is
+        called on cleanup.
+
+        If `C_Initialize()` reports that's it's already initialized, this is
+        treated as an error. Use this if you need perfect cleanup (ex: running
+        valgrind with --leak-check).
+        """
+
+    def __init__(self, *, file: str, behavior: InitializeFinalizeBehavior = None):
+        super().__init__()
+        if behavior is None:
+            behavior = Pkcs11Lib.InitializeFinalizeBehavior.DEFAULT
+
+        self._binding = _awscrt.pkcs11_lib_new(file, behavior)
