@@ -6,14 +6,15 @@ All network operations in `awscrt.mqtt` are asynchronous.
 
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0.
-
 import _awscrt
 from concurrent.futures import Future
 from enum import IntEnum
+from inspect import signature
 from awscrt import NativeResource
 import awscrt.exceptions
 from awscrt.http import HttpProxyOptions, HttpRequest
 from awscrt.io import ClientBootstrap, ClientTlsContext, SocketOptions
+from dataclasses import dataclass
 
 
 class QoS(IntEnum):
@@ -126,11 +127,45 @@ class Will:
         self.retain = retain
 
 
+@dataclass
+class OnConnectionSuccessData:
+    """Dataclass containing data related to a on_connection_success Callback
+
+    Args:
+        return_code (ConnectReturnCode): Connect return. code received from the server.
+        session_present (bool): True if the connection resumes an existing session.
+                                False if new session. Note that the server has forgotten all previous subscriptions
+                                if this is False.
+                                Subscriptions can be re-established via resubscribe_existing_topics() if the connection was a reconnection.
+    """
+    return_code: ConnectReturnCode = None
+    session_present: bool = False
+
+
+@dataclass
+class OnConnectionFailureData:
+    """Dataclass containing data related to a on_connection_failure Callback
+
+    Args:
+        error (ConnectReturnCode): Error code with reason for connection failure
+    """
+    error: awscrt.exceptions.AwsCrtError = None
+
+
+@dataclass
+class OnConnectionClosedData:
+    """Dataclass containing data related to a on_connection_closed Callback.
+    Currently unused.
+    """
+    pass
+
+
 class Client(NativeResource):
     """MQTT client.
 
     Args:
-        bootstrap (ClientBootstrap): Client bootstrap to use when initiating new socket connections.
+        bootstrap (Optional [ClientBootstrap]): Client bootstrap to use when initiating new socket connections.
+            If None is provided, the default singleton is used.
 
         tls_ctx (Optional[ClientTlsContext]): TLS context for secure socket connections.
             If None is provided, then an unencrypted connection is used.
@@ -138,13 +173,31 @@ class Client(NativeResource):
 
     __slots__ = ('tls_ctx')
 
-    def __init__(self, bootstrap, tls_ctx=None):
-        assert isinstance(bootstrap, ClientBootstrap)
+    def __init__(self, bootstrap=None, tls_ctx=None):
+        assert isinstance(bootstrap, ClientBootstrap) or bootstrap is None
         assert tls_ctx is None or isinstance(tls_ctx, ClientTlsContext)
 
         super().__init__()
         self.tls_ctx = tls_ctx
+        if not bootstrap:
+            bootstrap = ClientBootstrap.get_or_create_static_default()
         self._binding = _awscrt.mqtt_client_new(bootstrap, tls_ctx)
+
+
+@dataclass
+class OperationStatisticsData:
+    """Dataclass containing some simple statistics about the current state of the connection's queue of operations
+
+    Args:
+        incomplete_operation_count (int): total number of operations submitted to the connection that have not yet been completed.  Unacked operations are a subset of this.
+        incomplete_operation_size (int): total packet size of operations submitted to the connection that have not yet been completed.  Unacked operations are a subset of this.
+        unacked_operation_count (int): total number of operations that have been sent to the server and are waiting for a corresponding ACK before they can be completed.
+        unacked_operation_size (int): total packet size of operations that have been sent to the server and are waiting for a corresponding ACK before they can be completed.
+    """
+    incomplete_operation_count: int = 0
+    incomplete_operation_size: int = 0
+    unacked_operation_count: int = 0
+    unacked_operation_size: int = 0
 
 
 class Connection(NativeResource):
@@ -173,25 +226,50 @@ class Connection(NativeResource):
             The MQTT client will automatically attempt to reconnect.
             The function should take the following arguments return nothing:
 
-            *   `connection` (:class:`Connection`): This MQTT Connection.
+                *   `connection` (:class:`Connection`): This MQTT Connection.
 
-            *   `error` (:class:`awscrt.exceptions.AwsCrtError`): Exception which caused connection loss.
+                *   `error` (:class:`awscrt.exceptions.AwsCrtError`): Exception which caused connection loss.
 
-            *   `**kwargs` (dict): Forward-compatibility kwargs.
+                *   `**kwargs` (dict): Forward-compatibility kwargs.
 
         on_connection_resumed: Optional callback invoked whenever the MQTT connection
             is automatically resumed. Function should take the following arguments and return nothing:
 
-            *   `connection` (:class:`Connection`): This MQTT Connection
+                *   `connection` (:class:`Connection`): This MQTT Connection
 
-            *   `return_code` (:class:`ConnectReturnCode`): Connect return
-                code received from the server.
+                *   `return_code` (:class:`ConnectReturnCode`): Connect return
+                    code received from the server.
 
-            *   `session_present` (bool): True if resuming existing session. False if new session.
-                Note that the server has forgotten all previous subscriptions if this is False.
-                Subscriptions can be re-established via resubscribe_existing_topics().
+                *   `session_present` (bool): True if resuming existing session. False if new session.
+                    Note that the server has forgotten all previous subscriptions if this is False.
+                    Subscriptions can be re-established via resubscribe_existing_topics().
 
-            *   `**kwargs` (dict): Forward-compatibility kwargs.
+                *   `**kwargs` (dict): Forward-compatibility kwargs.
+
+        on_connection_success: Optional callback invoked whenever the connection successfully connects.
+            This callback is invoked for every successful connect and every successful reconnect.
+
+            Function should take the following arguments and return nothing:
+
+                * `connection` (:class:`Connection`): This MQTT Connection
+
+                * `callback_data` (:class:`OnConnectionSuccessData`): The data returned from the connection success.
+
+        on_connection_failure: Optional callback invoked whenever the connection fails to connect.
+            This callback is invoked for every failed connect and every failed reconnect.
+
+            Function should take the following arguments and return nothing:
+
+                * `connection` (:class:`Connection`): This MQTT Connection
+
+                * `callback_data` (:class:`OnConnectionFailureData`): The data returned from the connection failure.
+
+        on_connection_closed: Optional callback invoked whenever the connection has been disconnected and shutdown successfully.
+            Function should take the following arguments and return nothing:
+
+                * `connection` (:class:`Connection`): This MQTT Connection
+
+                * `callback_data` (:class:`OnConnectionClosedData`): The data returned from the connection close.
 
         reconnect_min_timeout_secs (int): Minimum time to wait between reconnect attempts.
             Must be <= `reconnect_max_timeout_secs`.
@@ -208,10 +286,13 @@ class Connection(NativeResource):
 
         ping_timeout_ms (int): Milliseconds to wait for ping response before client assumes
             the connection is invalid and attempts to reconnect.
-            This duration must be shorter than keep_alive_secs.
-            Alternatively, TCP keep-alive via :attr:`SocketOptions.keep_alive`
-            may accomplish this in a more efficient (low-power) scenario,
-            but keep-alive options may not work the same way on every platform and OS version.
+            This duration must be shorter than `keep_alive_secs`.
+
+        protocol_operation_timeout_ms (int): Milliseconds to wait for the response to the operation
+            requires response by protocol. Set to zero to disable timeout. Otherwise,
+            the operation will fail if no response is received within this amount of time after
+            the packet is written to the socket
+            It applied to PUBLISH (QoS>0) and UNSUBSCRIBE now.
 
         will (Will): Will to send with CONNECT packet. The will is
             published by the server when its connection to the client is unexpectedly lost.
@@ -225,7 +306,7 @@ class Connection(NativeResource):
         use_websocket (bool): If true, connect to MQTT over websockets.
 
         websocket_proxy_options (Optional[awscrt.http.HttpProxyOptions]):
-            Optional proxy options for websocket connections.
+            Optional proxy options for websocket connections.  Deprecated, use `proxy_options` instead.
 
         websocket_handshake_transform: Optional function to transform websocket handshake request.
             If provided, function is called each time a websocket connection is attempted.
@@ -233,11 +314,14 @@ class Connection(NativeResource):
             See :class:`WebsocketHandshakeTransformArgs` for more info.
             Function should take the following arguments and return nothing:
 
-            *   `transform_args` (:class:`WebsocketHandshakeTransformArgs`):
-                Contains HTTP request to be transformed. Function must call
-                `transform_args.done()` when complete.
+                *   `transform_args` (:class:`WebsocketHandshakeTransformArgs`):
+                    Contains HTTP request to be transformed. Function must call
+                    `transform_args.done()` when complete.
 
-            *   `**kwargs` (dict): Forward-compatibility kwargs.
+                *   `**kwargs` (dict): Forward-compatibility kwargs.
+
+        proxy_options (Optional[awscrt.http.HttpProxyOptions]):
+            Optional proxy options for all connections.
         """
 
     def __init__(self,
@@ -252,6 +336,7 @@ class Connection(NativeResource):
                  reconnect_max_timeout_secs=60,
                  keep_alive_secs=1200,
                  ping_timeout_ms=3000,
+                 protocol_operation_timeout_ms=0,
                  will=None,
                  username=None,
                  password=None,
@@ -259,6 +344,10 @@ class Connection(NativeResource):
                  use_websockets=False,
                  websocket_proxy_options=None,
                  websocket_handshake_transform=None,
+                 proxy_options=None,
+                 on_connection_success=None,
+                 on_connection_failure=None,
+                 on_connection_closed=None
                  ):
 
         assert isinstance(client, Client)
@@ -267,13 +356,21 @@ class Connection(NativeResource):
         assert isinstance(will, Will) or will is None
         assert isinstance(socket_options, SocketOptions) or socket_options is None
         assert isinstance(websocket_proxy_options, HttpProxyOptions) or websocket_proxy_options is None
+        assert isinstance(proxy_options, HttpProxyOptions) or proxy_options is None
         assert callable(websocket_handshake_transform) or websocket_handshake_transform is None
+        assert callable(on_connection_success) or on_connection_success is None
+        assert callable(on_connection_failure) or on_connection_failure is None
+        assert callable(on_connection_closed) or on_connection_closed is None
 
         if reconnect_min_timeout_secs > reconnect_max_timeout_secs:
             raise ValueError("'reconnect_min_timeout_secs' cannot exceed 'reconnect_max_timeout_secs'")
 
         if keep_alive_secs * 1000 <= ping_timeout_ms:
             raise ValueError("'keep_alive_secs' duration must be longer than 'ping_timeout_ms'")
+
+        if proxy_options and websocket_proxy_options:
+            raise ValueError("'websocket_proxy_options' has been deprecated in favor of 'proxy_options'.  "
+                             "Both parameters may not be set.")
 
         super().__init__()
 
@@ -283,6 +380,9 @@ class Connection(NativeResource):
         self._on_connection_resumed_cb = on_connection_resumed
         self._use_websockets = use_websockets
         self._ws_handshake_transform_cb = websocket_handshake_transform
+        self._on_connection_success_cb = on_connection_success
+        self._on_connection_failure_cb = on_connection_failure
+        self._on_connection_closed_cb = on_connection_closed
 
         # may be changed at runtime, take effect the the next time connect/reconnect occurs
         self.client_id = client_id
@@ -293,17 +393,33 @@ class Connection(NativeResource):
         self.reconnect_max_timeout_secs = reconnect_max_timeout_secs
         self.keep_alive_secs = keep_alive_secs
         self.ping_timeout_ms = ping_timeout_ms
+        self.protocol_operation_timeout_ms = protocol_operation_timeout_ms
         self.will = will
         self.username = username
         self.password = password
         self.socket_options = socket_options if socket_options else SocketOptions()
-        self.websocket_proxy_options = websocket_proxy_options
+        self.proxy_options = proxy_options if proxy_options else websocket_proxy_options
 
         self._binding = _awscrt.mqtt_client_connection_new(
             self,
             client,
             use_websockets,
         )
+
+    def _check_uses_old_message_callback_signature(self, callback):
+        # The callback used to have fewer args. Passing only those args, if it
+        # only has two args and no forward-compatibility to cover case where
+        # user function failed to take forward-compatibility **kwargs.
+
+        callback_sig = signature(callback)
+        try:
+            # try new signature
+            callback_sig.bind(topic='topic', payload='payload', dup=True, qos=QoS(1), retain=True)
+            return False
+        except TypeError:
+            # try old signature
+            callback_sig.bind(topic='topic', payload='payload')
+            return True
 
     def _on_connection_interrupted(self, error_code):
         if self._on_connection_interrupted_cb:
@@ -335,6 +451,26 @@ class Connection(NativeResource):
             # there's a chance the callback wasn't callable and user has no idea we tried to hand them the baton.
             if not future.done():
                 transform_args.set_done(e)
+
+    def _on_connection_closed(self):
+        if self:
+            if self._on_connection_closed_cb:
+                data = OnConnectionClosedData()
+                self._on_connection_closed_cb(connection=self, callback_data=data)
+
+    def _on_connection_success(self, return_code, session_present):
+        if self:
+            if self._on_connection_success_cb:
+                data = OnConnectionSuccessData(
+                    return_code=ConnectReturnCode(return_code),
+                    session_present=session_present)
+                self._on_connection_success_cb(connection=self, callback_data=data)
+
+    def _on_connection_failure(self, error_code):
+        if self:
+            if self._on_connection_failure_cb:
+                data = OnConnectionFailureData(error=awscrt.exceptions.from_code(error_code))
+                self._on_connection_failure_cb(connection=self, callback_data=data)
 
     def connect(self):
         """Open the actual connection to the server (async).
@@ -368,12 +504,13 @@ class Connection(NativeResource):
                 self.reconnect_max_timeout_secs,
                 self.keep_alive_secs,
                 self.ping_timeout_ms,
+                self.protocol_operation_timeout_ms,
                 self.will,
                 self.username,
                 self.password,
                 self.clean_session,
                 on_connect,
-                self.websocket_proxy_options
+                self.proxy_options
             )
 
         except Exception as e:
@@ -420,6 +557,7 @@ class Connection(NativeResource):
 
         try:
             _awscrt.mqtt_client_connection_disconnect(self._binding, on_disconnect)
+
         except Exception as e:
             future.set_exception(e)
 
@@ -444,11 +582,19 @@ class Connection(NativeResource):
             callback: Optional callback invoked when message received.
                 Function should take the following arguments and return nothing:
 
-                *   `topic` (str): Topic receiving message.
+                    *   `topic` (str): Topic receiving message.
 
-                *   `payload` (bytes): Payload of message.
+                    *   `payload` (bytes): Payload of message.
 
-                *   `**kwargs` (dict): Forward-compatibility kwargs.
+                    *   `dup` (bool): DUP flag. If True, this might be re-delivery
+                        of an earlier attempt to send the message.
+
+                    *   `qos` (:class:`QoS`): Quality of Service used to deliver the message.
+
+                    *   `retain` (bool): Retain flag. If True, the message was sent
+                        as a result of a new subscription being made by the client.
+
+                    *   `**kwargs` (dict): Forward-compatibility kwargs.
 
         Returns:
             Tuple[concurrent.futures.Future, int]: Tuple containing a Future and
@@ -456,12 +602,12 @@ class Connection(NativeResource):
             SUBACK is received from the server. If successful, the Future will
             contain a dict with the following members:
 
-            *   ['packet_id'] (int): ID of the SUBSCRIBE packet being acknowledged.
+                *   ['packet_id'] (int): ID of the SUBSCRIBE packet being acknowledged.
 
-            *   ['topic'] (str): Topic filter of the SUBSCRIBE packet being acknowledged.
+                *   ['topic'] (str): Topic filter of the SUBSCRIBE packet being acknowledged.
 
-            *   ['qos'] (:class:`QoS`): Maximum QoS that was granted by the server.
-                This may be lower than the requested QoS.
+                *   ['qos'] (:class:`QoS`): Maximum QoS that was granted by the server.
+                    This may be lower than the requested QoS.
 
             If unsuccessful, the Future contains an exception. The exception
             will be a :class:`SubscribeError` if a SUBACK was received
@@ -473,8 +619,14 @@ class Connection(NativeResource):
         packet_id = 0
 
         if callback:
-            def callback_wrapper(topic, payload):
-                callback(topic=topic, payload=payload)
+            uses_old_signature = self._check_uses_old_message_callback_signature(callback)
+
+            def callback_wrapper(topic, payload, dup, qos, retain):
+                if uses_old_signature:
+                    callback(topic=topic, payload=payload)
+                else:
+                    callback(topic=topic, payload=payload, dup=dup, qos=QoS(qos), retain=retain)
+
         else:
             callback_wrapper = None
 
@@ -508,17 +660,32 @@ class Connection(NativeResource):
         callback: Callback to invoke when message received, or None to disable.
             Function should take the following arguments and return nothing:
 
-            *   `topic` (str): Topic receiving message.
+                *   `topic` (str): Topic receiving message.
 
-            *   `payload` (bytes): Payload of message.
+                *   `payload` (bytes): Payload of message.
 
-            *   `**kwargs` (dict): Forward-compatibility kwargs.
+                *   `dup` (bool): DUP flag. If True, this might be re-delivery
+                    of an earlier attempt to send the message.
+
+                *   `qos` (:class:`QoS`): Quality of Service used to deliver the message.
+
+                *   `retain` (bool): Retain flag. If True, the message was sent
+                    as a result of a new subscription being made by the client.
+
+                *   `**kwargs` (dict): Forward-compatibility kwargs.
         """
         assert callable(callback) or callback is None
 
         if callback:
-            def callback_wrapper(topic, payload):
-                callback(topic=topic, payload=payload)
+
+            uses_old_signature = self._check_uses_old_message_callback_signature(callback)
+
+            def callback_wrapper(topic, payload, dup, qos, retain):
+                if uses_old_signature:
+                    callback(topic=topic, payload=payload)
+                else:
+                    callback(topic=topic, payload=payload, dup=dup, qos=QoS(qos), retain=retain)
+
         else:
             callback_wrapper = None
 
@@ -563,6 +730,15 @@ class Connection(NativeResource):
 
         This is to help when resuming a connection with a clean session.
 
+        **Important**: Currently the resubscribe function does not take the AWS IoT Core maximum subscriptions
+        per subscribe request quota into account. If the client has more subscriptions than the maximum,
+        resubscribing must be done manually using the `subscribe()` function for each desired topic
+        filter. The client will be disconnected by AWS IoT Core if the resubscribe exceeds the subscriptions
+        per subscribe request quota.
+
+        The AWS IoT Core maximum subscriptions per subscribe request quota is listed at the following URL:
+        https://docs.aws.amazon.com/general/latest/gr/iot-core.html#genref_max_subscriptions_per_subscribe_request
+
         Returns:
             Tuple[concurrent.futures.Future, int]: Tuple containing a Future and
             the ID of the SUBSCRIBE packet. The Future completes when a SUBACK
@@ -606,7 +782,7 @@ class Connection(NativeResource):
 
         Args:
             topic (str): Topic name.
-            payload (buffer): Contents of message.
+            payload (Union[str, bytes, bytearray]): Contents of message.
             qos (QoS): Quality of Service for delivering this message.
             retain (bool): If True, the server will store the message and its QoS
                 so that it can be delivered to future subscribers whose subscriptions
@@ -639,6 +815,16 @@ class Connection(NativeResource):
             future.set_exception(e)
 
         return future, packet_id
+
+    def get_stats(self):
+        """Queries the connection's internal statistics for incomplete operations.
+
+        Returns:
+            The (:class:`OperationStatisticsData`) containing the statistics
+        """
+
+        result = _awscrt.mqtt_client_connection_get_stats(self._binding)
+        return OperationStatisticsData(result[0], result[1], result[2], result[3])
 
 
 class WebsocketHandshakeTransformArgs:
