@@ -4,6 +4,8 @@
  */
 #include "io.h"
 
+#include <aws/common/atomics.h>
+
 #include <aws/io/channel_bootstrap.h>
 #include <aws/io/event_loop.h>
 #include <aws/io/socket.h>
@@ -81,6 +83,18 @@ PyObject *aws_py_is_alpn_available(PyObject *self, PyObject *args) {
     return PyBool_FromLong(aws_tls_is_alpn_available());
 }
 
+PyObject *aws_py_is_tls_cipher_supported(PyObject *self, PyObject *args) {
+    (void)self;
+    (void)args;
+
+    int cipher_pref = 0;
+
+    if (!PyArg_ParseTuple(args, "i", &cipher_pref)) {
+        return NULL;
+    }
+
+    return PyBool_FromLong(aws_tls_is_cipher_pref_supported(cipher_pref));
+}
 /*******************************************************************************
  * AWS_EVENT_LOOP_GROUP
  ******************************************************************************/
@@ -131,8 +145,10 @@ PyObject *aws_py_event_loop_group_new(PyObject *self, PyObject *args) {
     struct aws_allocator *allocator = aws_py_get_allocator();
 
     uint16_t num_threads;
+    int is_pinned;
+    uint16_t cpu_group;
     PyObject *shutdown_complete_py;
-    if (!PyArg_ParseTuple(args, "HO", &num_threads, &shutdown_complete_py)) {
+    if (!PyArg_ParseTuple(args, "HpHO", &num_threads, &is_pinned, &cpu_group, &shutdown_complete_py)) {
         return NULL;
     }
 
@@ -146,7 +162,12 @@ PyObject *aws_py_event_loop_group_new(PyObject *self, PyObject *args) {
         .shutdown_callback_user_data = binding,
     };
 
-    binding->native = aws_event_loop_group_new_default(allocator, num_threads, &shutdown_options);
+    if (is_pinned) {
+        binding->native =
+            aws_event_loop_group_new_default_pinned_to_cpu_group(allocator, num_threads, cpu_group, &shutdown_options);
+    } else {
+        binding->native = aws_event_loop_group_new_default(allocator, num_threads, &shutdown_options);
+    }
     if (binding->native == NULL) {
         PyErr_SetAwsLastError();
         goto elg_init_failed;
@@ -222,8 +243,12 @@ PyObject *aws_py_host_resolver_new_default(PyObject *self, PyObject *args) {
     }
 
     /* From hereon, we need to clean up if errors occur */
+    struct aws_host_resolver_default_options resolver_options = {
+        .max_entries = max_hosts,
+        .el_group = elg,
+    };
 
-    host_resolver->native = aws_host_resolver_new_default(allocator, max_hosts, elg, NULL);
+    host_resolver->native = aws_host_resolver_new_default(allocator, &resolver_options);
     if (host_resolver->native == NULL) {
         PyErr_SetAwsLastError();
         goto resolver_init_failed;
@@ -389,7 +414,8 @@ PyObject *aws_py_client_tls_ctx_new(PyObject *self, PyObject *args) {
 
     struct aws_allocator *allocator = aws_py_get_allocator();
 
-    int min_tls_version;
+    int min_tls_version = 0;
+    int cipher_pref = 0;
     const char *ca_dirpath;
     const char *ca_buffer;
     Py_ssize_t ca_buffer_len;
@@ -400,49 +426,124 @@ PyObject *aws_py_client_tls_ctx_new(PyObject *self, PyObject *args) {
     Py_ssize_t private_key_buffer_len;
     const char *pkcs12_filepath;
     const char *pkcs12_password;
-    uint8_t verify_peer;
+    int verify_peer; /* p - boolean predicate */
+    PyObject *py_pkcs11_lib;
+    const char *pkcs11_user_pin;
+    Py_ssize_t pkcs11_user_pin_len;
+    PyObject *py_pkcs11_slot_id;
+    const char *pkcs11_token_label;
+    Py_ssize_t pkcs11_token_label_len;
+    const char *pkcs11_priv_key_label;
+    Py_ssize_t pkcs11_priv_key_label_len;
+    const char *pkcs11_cert_file_path;
+    Py_ssize_t pkcs11_cert_file_path_len;
+    const char *pkcs11_cert_file_contents;
+    Py_ssize_t pkcs11_cert_file_contents_len;
+    const char *windows_cert_store_path;
+
     if (!PyArg_ParseTuple(
             args,
-            "bzz#zz#z#zzb",
-            &min_tls_version,
-            &ca_dirpath,
-            &ca_buffer,
-            &ca_buffer_len,
-            &alpn_list,
-            &certificate_buffer,
-            &certificate_buffer_len,
-            &private_key_buffer,
-            &private_key_buffer_len,
-            &pkcs12_filepath,
-            &pkcs12_password,
-            &verify_peer)) {
+            "iizz#zz#z#zzpOz#Oz#z#z#z#z",
+            /* i */ &min_tls_version,
+            /* i */ &cipher_pref,
+            /* z */ &ca_dirpath,
+            /* z */ &ca_buffer,
+            /* # */ &ca_buffer_len,
+            /* z */ &alpn_list,
+            /* z */ &certificate_buffer,
+            /* # */ &certificate_buffer_len,
+            /* z */ &private_key_buffer,
+            /* # */ &private_key_buffer_len,
+            /* z */ &pkcs12_filepath,
+            /* z */ &pkcs12_password,
+            /* p */ &verify_peer,
+            /* O */ &py_pkcs11_lib,
+            /* z */ &pkcs11_user_pin,
+            /* # */ &pkcs11_user_pin_len,
+            /* O */ &py_pkcs11_slot_id,
+            /* z */ &pkcs11_token_label,
+            /* # */ &pkcs11_token_label_len,
+            /* z */ &pkcs11_priv_key_label,
+            /* # */ &pkcs11_priv_key_label_len,
+            /* z */ &pkcs11_cert_file_path,
+            /* # */ &pkcs11_cert_file_path_len,
+            /* z */ &pkcs11_cert_file_contents,
+            /* # */ &pkcs11_cert_file_contents_len,
+            /* z */ &windows_cert_store_path)) {
         return NULL;
     }
 
     struct aws_tls_ctx_options ctx_options;
     AWS_ZERO_STRUCT(ctx_options);
-    if (certificate_buffer_len > 0 && private_key_buffer_len > 0) {
+    if (certificate_buffer != NULL) {
+        /* mTLS with certificate and private key*/
         struct aws_byte_cursor cert = aws_byte_cursor_from_array(certificate_buffer, certificate_buffer_len);
         struct aws_byte_cursor key = aws_byte_cursor_from_array(private_key_buffer, private_key_buffer_len);
         if (aws_tls_ctx_options_init_client_mtls(&ctx_options, allocator, &cert, &key)) {
             PyErr_SetAwsLastError();
             return NULL;
         }
+    } else if (py_pkcs11_lib != Py_None) {
+        /* mTLS with PKCS#11 */
+        struct aws_pkcs11_lib *pkcs11_lib = aws_py_get_pkcs11_lib(py_pkcs11_lib);
+        if (pkcs11_lib == NULL) {
+            return NULL;
+        }
+
+        bool has_slot_id = false;
+        uint64_t slot_id_value = 0;
+        if (py_pkcs11_slot_id != Py_None) {
+            has_slot_id = true;
+            slot_id_value = PyLong_AsUnsignedLongLong(py_pkcs11_slot_id);
+            if ((slot_id_value == (uint64_t)-1) && PyErr_Occurred()) {
+                PyErr_SetString(PyExc_ValueError, "PKCS#11 slot_id is not a valid int");
+                return NULL;
+            }
+        }
+
+        struct aws_tls_ctx_pkcs11_options pkcs11_options = {
+            .pkcs11_lib = pkcs11_lib,
+            .user_pin = aws_byte_cursor_from_array(pkcs11_user_pin, pkcs11_user_pin_len),
+            .slot_id = has_slot_id ? &slot_id_value : NULL,
+            .token_label = aws_byte_cursor_from_array(pkcs11_token_label, pkcs11_token_label_len),
+            .private_key_object_label = aws_byte_cursor_from_array(pkcs11_priv_key_label, pkcs11_priv_key_label_len),
+            .cert_file_path = aws_byte_cursor_from_array(pkcs11_cert_file_path, pkcs11_cert_file_path_len),
+            .cert_file_contents = aws_byte_cursor_from_array(pkcs11_cert_file_contents, pkcs11_cert_file_contents_len),
+        };
+
+        if (aws_tls_ctx_options_init_client_mtls_with_pkcs11(&ctx_options, allocator, &pkcs11_options)) {
+            return PyErr_AwsLastError();
+        }
+    } else if (pkcs12_filepath != NULL) {
+        /* mTLS with PKCS#12 */
+        struct aws_byte_cursor password = aws_byte_cursor_from_c_str(pkcs12_password);
+        if (aws_tls_ctx_options_init_client_mtls_pkcs12_from_path(
+                &ctx_options, allocator, pkcs12_filepath, &password)) {
+            return PyErr_AwsLastError();
+        }
+    } else if (windows_cert_store_path != NULL) {
+        /* mTLS with certificate from a Windows certificate store */
+        if (aws_tls_ctx_options_init_client_mtls_from_system_path(&ctx_options, allocator, windows_cert_store_path)) {
+            return PyErr_AwsLastError();
+        }
     } else {
+        /* no mTLS */
         aws_tls_ctx_options_init_default_client(&ctx_options, allocator);
     }
 
     /* From hereon, we need to clean up if errors occur */
 
     ctx_options.minimum_tls_version = min_tls_version;
+    ctx_options.cipher_pref = cipher_pref;
 
-    if (ca_dirpath) {
+    if (ca_dirpath != NULL) {
         if (aws_tls_ctx_options_override_default_trust_store_from_path(&ctx_options, ca_dirpath, NULL)) {
             PyErr_SetAwsLastError();
             goto ctx_options_failure;
         }
     }
-    if (ca_buffer_len > 0) {
+
+    if (ca_buffer != NULL) {
         struct aws_byte_cursor ca = aws_byte_cursor_from_array(ca_buffer, ca_buffer_len);
 
         if (aws_tls_ctx_options_override_default_trust_store(&ctx_options, &ca)) {
@@ -451,23 +552,13 @@ PyObject *aws_py_client_tls_ctx_new(PyObject *self, PyObject *args) {
         }
     }
 
-    if (alpn_list) {
+    if (alpn_list != NULL) {
         if (aws_tls_ctx_options_set_alpn_list(&ctx_options, alpn_list)) {
             PyErr_SetAwsLastError();
             goto ctx_options_failure;
         }
     }
 
-#ifdef __APPLE__
-    if (pkcs12_filepath && pkcs12_password) {
-        struct aws_byte_cursor password = aws_byte_cursor_from_c_str(pkcs12_password);
-        if (aws_tls_ctx_options_init_client_mtls_pkcs12_from_path(
-                &ctx_options, allocator, pkcs12_filepath, &password)) {
-            PyErr_SetAwsLastError();
-            goto ctx_options_failure;
-        }
-    }
-#endif
     ctx_options.verify_peer = (bool)verify_peer;
     struct aws_tls_ctx *tls_ctx = aws_tls_client_ctx_new(allocator, &ctx_options);
     if (!tls_ctx) {
@@ -624,25 +715,23 @@ PyObject *aws_py_tls_connection_options_set_server_name(PyObject *self, PyObject
 /* aws_input_stream implementation for accessing Python I/O classes */
 struct aws_input_stream_py_impl {
     struct aws_input_stream base;
+    struct aws_allocator *allocator;
 
     bool is_end_of_stream;
 
-    /* Dependencies that must outlive this */
-    PyObject *io;
-};
+    /* Track the refcount from C land */
+    struct aws_atomic_var c_ref;
 
-static void s_aws_input_stream_py_destroy(struct aws_input_stream *stream) {
-    struct aws_input_stream_py_impl *impl = stream->impl;
-    Py_DECREF(impl->io);
-    aws_mem_release(stream->allocator, stream);
-}
+    /* Pointer to python self. The stream will have a same lifetime as the python Object */
+    PyObject *py_self;
+};
 
 static int s_aws_input_stream_py_seek(
     struct aws_input_stream *stream,
-    aws_off_t offset,
+    int64_t offset,
     enum aws_stream_seek_basis basis) {
 
-    struct aws_input_stream_py_impl *impl = stream->impl;
+    struct aws_input_stream_py_impl *impl = AWS_CONTAINER_OF(stream, struct aws_input_stream_py_impl, base);
 
     int aws_result = AWS_OP_SUCCESS;
     PyObject *method_result = NULL;
@@ -653,7 +742,7 @@ static int s_aws_input_stream_py_seek(
         return AWS_OP_ERR; /* Python has shut down. Nothing matters anymore, but don't crash */
     }
 
-    method_result = PyObject_CallMethod(impl->io, "seek", "(li)", offset, basis);
+    method_result = PyObject_CallMethod(impl->py_self, "_seek", "(Li)", offset, basis);
     if (!method_result) {
         aws_result = aws_py_raise_error();
         goto done;
@@ -671,7 +760,7 @@ done:
 }
 
 int s_aws_input_stream_py_read(struct aws_input_stream *stream, struct aws_byte_buf *dest) {
-    struct aws_input_stream_py_impl *impl = stream->impl;
+    struct aws_input_stream_py_impl *impl = AWS_CONTAINER_OF(stream, struct aws_input_stream_py_impl, base);
 
     int aws_result = AWS_OP_SUCCESS;
     PyObject *memory_view = NULL;
@@ -689,7 +778,7 @@ int s_aws_input_stream_py_read(struct aws_input_stream *stream, struct aws_byte_
         goto done;
     }
 
-    method_result = PyObject_CallMethod(impl->io, "readinto", "(O)", memory_view);
+    method_result = PyObject_CallMethod(impl->py_self, "_read_into_memoryview", "(O)", memory_view);
     if (!method_result) {
         aws_result = aws_py_raise_error();
         goto done;
@@ -723,7 +812,7 @@ done:
 }
 
 int s_aws_input_stream_py_get_status(struct aws_input_stream *stream, struct aws_stream_status *status) {
-    struct aws_input_stream_py_impl *impl = stream->impl;
+    struct aws_input_stream_py_impl *impl = AWS_CONTAINER_OF(stream, struct aws_input_stream_py_impl, base);
 
     status->is_valid = true;
     status->is_end_of_stream = impl->is_end_of_stream;
@@ -734,7 +823,39 @@ int s_aws_input_stream_py_get_status(struct aws_input_stream *stream, struct aws
 int s_aws_input_stream_py_get_length(struct aws_input_stream *stream, int64_t *out_length) {
     (void)stream;
     (void)out_length;
-    return AWS_ERROR_UNIMPLEMENTED;
+    return aws_raise_error(AWS_ERROR_UNIMPLEMENTED);
+}
+
+void s_aws_input_stream_py_acquire(struct aws_input_stream *stream) {
+    struct aws_input_stream_py_impl *impl = AWS_CONTAINER_OF(stream, struct aws_input_stream_py_impl, base);
+    size_t pre_ref = aws_atomic_fetch_add(&impl->c_ref, 1);
+    if (pre_ref == 0) {
+        /* Only acquire the python ref when it's a new C ref */
+        /*************** GIL ACQUIRE ***************/
+        PyGILState_STATE state;
+        if (aws_py_gilstate_ensure(&state)) {
+            return; /* Python has shut down. Nothing matters anymore, but don't crash */
+        }
+        Py_INCREF(impl->py_self);
+        PyGILState_Release(state);
+        /*************** GIL RELEASE ***************/
+    }
+}
+
+void s_aws_input_stream_py_release(struct aws_input_stream *stream) {
+    struct aws_input_stream_py_impl *impl = AWS_CONTAINER_OF(stream, struct aws_input_stream_py_impl, base);
+    size_t pre_ref = aws_atomic_fetch_sub(&impl->c_ref, 1);
+    if (pre_ref == 1) {
+        /* Only release the python ref when all the C refs gone */
+        /*************** GIL ACQUIRE ***************/
+        PyGILState_STATE state;
+        if (aws_py_gilstate_ensure(&state)) {
+            return; /* Python has shut down. Nothing matters anymore, but don't crash */
+        }
+        Py_DECREF(impl->py_self);
+        PyGILState_Release(state);
+        /*************** GIL RELEASE ***************/
+    }
 }
 
 static struct aws_input_stream_vtable s_aws_input_stream_py_vtable = {
@@ -742,30 +863,9 @@ static struct aws_input_stream_vtable s_aws_input_stream_py_vtable = {
     .read = s_aws_input_stream_py_read,
     .get_status = s_aws_input_stream_py_get_status,
     .get_length = s_aws_input_stream_py_get_length,
-    .destroy = s_aws_input_stream_py_destroy,
+    .acquire = s_aws_input_stream_py_acquire,
+    .release = s_aws_input_stream_py_release,
 };
-
-static struct aws_input_stream *aws_input_stream_new_from_py(PyObject *io) {
-
-    if (!io || (io == Py_None)) {
-        aws_raise_error(AWS_ERROR_INVALID_ARGUMENT);
-        return NULL;
-    }
-
-    struct aws_allocator *alloc = aws_py_get_allocator();
-    struct aws_input_stream_py_impl *impl = aws_mem_calloc(alloc, 1, sizeof(struct aws_input_stream_py_impl));
-    if (!impl) {
-        return NULL;
-    }
-
-    impl->base.allocator = alloc;
-    impl->base.vtable = &s_aws_input_stream_py_vtable;
-    impl->base.impl = impl;
-    impl->io = io;
-    Py_INCREF(impl->io);
-
-    return &impl->base;
-}
 
 /**
  * Begin aws_input_stream <--> InputStream binding code.
@@ -777,26 +877,34 @@ static struct aws_input_stream *aws_input_stream_new_from_py(PyObject *io) {
 
 static void s_input_stream_capsule_destructor(PyObject *py_capsule) {
     struct aws_input_stream *stream = PyCapsule_GetPointer(py_capsule, s_capsule_name_input_stream);
-    aws_input_stream_destroy(stream);
+    struct aws_input_stream_py_impl *impl = AWS_CONTAINER_OF(stream, struct aws_input_stream_py_impl, base);
+    aws_mem_release(impl->allocator, impl);
 }
 
 PyObject *aws_py_input_stream_new(PyObject *self, PyObject *args) {
     (void)self;
 
-    PyObject *py_io;
-    if (!PyArg_ParseTuple(args, "O", &py_io)) {
+    PyObject *py_self;
+    if (!PyArg_ParseTuple(args, "O", &py_self)) {
         return NULL;
     }
 
-    struct aws_input_stream *stream = aws_input_stream_new_from_py(py_io);
-    if (!stream) {
-        return PyErr_AwsLastError();
+    if (py_self == Py_None) {
+        PyErr_SetString(PyExc_TypeError, "InputStream cannot be None");
+        return NULL;
     }
 
-    PyObject *py_capsule = PyCapsule_New(stream, s_capsule_name_input_stream, s_input_stream_capsule_destructor);
+    struct aws_allocator *alloc = aws_py_get_allocator();
+    struct aws_input_stream_py_impl *impl = aws_mem_calloc(alloc, 1, sizeof(struct aws_input_stream_py_impl));
+    impl->allocator = alloc;
+    impl->base.vtable = &s_aws_input_stream_py_vtable;
+    impl->py_self = py_self;
+    aws_atomic_init_int(&impl->c_ref, 0);
+    /* Lifetime of the impl will be the same as py_capsule and being handled by python */
+    PyObject *py_capsule = PyCapsule_New(&impl->base, s_capsule_name_input_stream, s_input_stream_capsule_destructor);
+
     if (!py_capsule) {
-        aws_input_stream_destroy(stream);
-        return NULL;
+        aws_mem_release(impl->allocator, impl);
     }
 
     return py_capsule;
