@@ -1,7 +1,6 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0.
 
-from __future__ import absolute_import, print_function
 import awscrt.http
 import awscrt.io
 import enum
@@ -22,7 +21,6 @@ Stage = enum.Enum('Stage', [
     'HttpClientConnected',
     'HttpStreamStart',
     'HttpStreamReceivingBody',
-    'HttpStreamReceivingBodyMainThread',
     'HttpStreamDone',
     'HttpConnectionClose',
     'HttpConnectionDone',
@@ -83,7 +81,13 @@ if __name__ == '__main__':
     set_stage(Stage.HttpClientStart)
 
     # HttpClientConnected
-    http_connection = http_connection.result(TIMEOUT)
+    try:
+        http_connection = http_connection.result(TIMEOUT)
+    except Exception:
+        # the internet's a flaky place and this isn't a correctness test
+        print("Connection failed. Exiting out early...")
+        set_stage(Stage.Done)
+
     set_stage(Stage.HttpClientConnected)
 
     # HttpStreamStart
@@ -93,21 +97,23 @@ if __name__ == '__main__':
     receiving_body_event = threading.Event()
 
     def on_incoming_body(http_stream, chunk):
-        # HttpStreamReceivingBody: Exit from event-loop thread while receiving body
-        set_stage(Stage.HttpStreamReceivingBody)
         receiving_body_event.set()
 
     http_stream = http_connection.request(request, on_body=on_incoming_body)
     http_stream.activate()
     set_stage(Stage.HttpStreamStart)
 
-    # HttpStreamReceivingBodyMainThread: Exit from main thread while receiving body
+    # HttpStreamReceivingBody
     assert receiving_body_event.wait(TIMEOUT)
-    set_stage(Stage.HttpStreamReceivingBodyMainThread)
+    set_stage(Stage.HttpStreamReceivingBody)
 
     # HttpStreamDone
-    status_code = http_stream.completion_future.result(TIMEOUT)
-    assert(status_code == 200)
+    try:
+        status_code = http_stream.completion_future.result(TIMEOUT)
+    except Exception:
+        # the internet's a flaky place and this isn't a correctness test
+        print("Request failed. Continuing with cleanup...")
+
     del http_stream
     del request
     set_stage(Stage.HttpStreamDone)
