@@ -1,18 +1,20 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0.
 
-from __future__ import absolute_import
 import awscrt.auth
 import awscrt.io
 import datetime
 from io import BytesIO
 import os
+import sys
 from test import NativeResourceTest, TIMEOUT
 import time
+import unittest
 
 EXAMPLE_ACCESS_KEY_ID = 'example_access_key_id'
 EXAMPLE_SECRET_ACCESS_KEY = 'example_secret_access_key'
 EXAMPLE_SESSION_TOKEN = 'example_session_token'
+EXAMPLE_SESSION_EXPIRATION = datetime.datetime.fromtimestamp(1609911816, tz=datetime.timezone.utc)
 
 
 class ScopedEnvironmentVariable:
@@ -26,11 +28,14 @@ class ScopedEnvironmentVariable:
         self.prev_value = os.environ.get(key)
 
     def __enter__(self):
-        os.environ[self.key] = self.value
+        if self.value is None:
+            os.environ.pop(self.key, None)
+        else:
+            os.environ[self.key] = self.value
 
     def __exit__(self, type, value, tb):
         if self.prev_value is None:
-            del os.environ[self.key]
+            os.environ.pop(self.key, None)
         else:
             os.environ[self.key] = self.prev_value
 
@@ -40,12 +45,14 @@ class TestCredentials(NativeResourceTest):
         credentials = awscrt.auth.AwsCredentials(
             EXAMPLE_ACCESS_KEY_ID,
             EXAMPLE_SECRET_ACCESS_KEY,
-            EXAMPLE_SESSION_TOKEN)
+            EXAMPLE_SESSION_TOKEN,
+            EXAMPLE_SESSION_EXPIRATION)
 
         # Don't use assertEqual(), which could log actual credentials if test fails.
         self.assertTrue(EXAMPLE_ACCESS_KEY_ID == credentials.access_key_id)
         self.assertTrue(EXAMPLE_SECRET_ACCESS_KEY == credentials.secret_access_key)
         self.assertTrue(EXAMPLE_SESSION_TOKEN == credentials.session_token)
+        self.assertTrue(EXAMPLE_SESSION_EXPIRATION == credentials.expiration)
 
     def test_create_no_session_token(self):
         credentials = awscrt.auth.AwsCredentials(EXAMPLE_ACCESS_KEY_ID, EXAMPLE_SECRET_ACCESS_KEY)
@@ -54,6 +61,14 @@ class TestCredentials(NativeResourceTest):
         self.assertTrue(EXAMPLE_ACCESS_KEY_ID == credentials.access_key_id)
         self.assertTrue(EXAMPLE_SECRET_ACCESS_KEY == credentials.secret_access_key)
         self.assertTrue(credentials.session_token is None)
+
+    def test_create_no_expiration(self):
+        credentials = awscrt.auth.AwsCredentials(EXAMPLE_ACCESS_KEY_ID, EXAMPLE_SECRET_ACCESS_KEY)
+
+        # Don't use assertEqual(), which could log actual credentials if test fails.
+        self.assertTrue(EXAMPLE_ACCESS_KEY_ID == credentials.access_key_id)
+        self.assertTrue(EXAMPLE_SECRET_ACCESS_KEY == credentials.secret_access_key)
+        self.assertTrue(credentials.expiration is None)
 
 
 class TestProvider(NativeResourceTest):
@@ -87,7 +102,8 @@ class TestProvider(NativeResourceTest):
     def test_default_provider(self):
         # Default credentials provider should pick up environment variables.
         with ScopedEnvironmentVariable('AWS_ACCESS_KEY_ID', EXAMPLE_ACCESS_KEY_ID), \
-                ScopedEnvironmentVariable('AWS_SECRET_ACCESS_KEY', EXAMPLE_SECRET_ACCESS_KEY):
+                ScopedEnvironmentVariable('AWS_SECRET_ACCESS_KEY', EXAMPLE_SECRET_ACCESS_KEY),\
+                ScopedEnvironmentVariable('AWS_SESSION_TOKEN', None):
 
             event_loop_group = awscrt.io.EventLoopGroup()
             host_resolver = awscrt.io.DefaultHostResolver(event_loop_group)
@@ -101,6 +117,188 @@ class TestProvider(NativeResourceTest):
             self.assertTrue(EXAMPLE_ACCESS_KEY_ID == credentials.access_key_id)
             self.assertTrue(EXAMPLE_SECRET_ACCESS_KEY == credentials.secret_access_key)
             self.assertTrue(credentials.session_token is None)
+
+    def test_profile_provider(self):
+        # Profile provider should pick up the profile file to provide the credentials.
+        profile_name = "crt_user"
+        credentials_filepath = "test/resources/example_profile"
+        event_loop_group = awscrt.io.EventLoopGroup()
+        host_resolver = awscrt.io.DefaultHostResolver(event_loop_group)
+        bootstrap = awscrt.io.ClientBootstrap(event_loop_group, host_resolver)
+        provider = awscrt.auth.AwsCredentialsProvider.new_profile(
+            bootstrap, profile_name=profile_name, credentials_filepath=credentials_filepath)
+
+        future = provider.get_credentials()
+        credentials = future.result(TIMEOUT)
+
+        # Don't use assertEqual(), which could log actual credentials if test fails.
+        self.assertTrue(EXAMPLE_ACCESS_KEY_ID == credentials.access_key_id)
+        self.assertTrue(EXAMPLE_SECRET_ACCESS_KEY == credentials.secret_access_key)
+        self.assertTrue(credentials.session_token is None)
+
+    def test_environment_provider(self):
+        with ScopedEnvironmentVariable('AWS_ACCESS_KEY_ID', EXAMPLE_ACCESS_KEY_ID), \
+                ScopedEnvironmentVariable('AWS_SECRET_ACCESS_KEY', EXAMPLE_SECRET_ACCESS_KEY), \
+                ScopedEnvironmentVariable('AWS_SESSION_TOKEN', EXAMPLE_SESSION_TOKEN):
+
+            provider = awscrt.auth.AwsCredentialsProvider.new_environment()
+            credentials = provider.get_credentials().result(TIMEOUT)
+
+            # Don't use assertEqual(), which could log actual credentials if test fails.
+            self.assertTrue(EXAMPLE_ACCESS_KEY_ID == credentials.access_key_id)
+            self.assertTrue(EXAMPLE_SECRET_ACCESS_KEY == credentials.secret_access_key)
+            self.assertTrue(EXAMPLE_SESSION_TOKEN == credentials.session_token)
+
+    def test_chain_provider(self):
+        provider = awscrt.auth.AwsCredentialsProvider.new_chain([
+            awscrt.auth.AwsCredentialsProvider.new_static('id_a', 'secret_a'),
+            awscrt.auth.AwsCredentialsProvider.new_static('id_b', 'secret_b'),
+        ])
+        credentials = provider.get_credentials().result(TIMEOUT)
+        self.assertTrue('id_a' == credentials.access_key_id)
+        self.assertTrue('secret_a' == credentials.secret_access_key)
+        self.assertTrue(credentials.session_token is None)
+
+    def test_chain_provider_bad_args(self):
+        with self.assertRaises(TypeError):
+            awscrt.auth.AwsCredentialsProvider.new_chain(None)
+
+        with self.assertRaises(ValueError):
+            awscrt.auth.AwsCredentialsProvider.new_chain([])
+
+        with self.assertRaises(TypeError):
+            provider = awscrt.auth.AwsCredentialsProvider.new_chain([
+                awscrt.auth.AwsCredentialsProvider.new_static('id_a', 'secret_a'),
+                "I am not an AwsCredentialsProvider",
+            ])
+
+    def test_process_provider(self):
+        with ScopedEnvironmentVariable("AWS_CONFIG_FILE", "test/resources/example_config"):
+            if sys.platform == 'win32':
+                profile = 'test_process_provider_win'
+            else:
+                profile = 'test_process_provider'
+            provider = awscrt.auth.AwsCredentialsProvider.new_process(profile)
+            credentials = provider.get_credentials().result(TIMEOUT)
+
+            # Don't use assertEqual(), which could log actual credentials if test fails.
+            self.assertTrue('process_access_key_id' == credentials.access_key_id)
+            self.assertTrue('process_secret_access_key' == credentials.secret_access_key)
+            self.assertTrue(credentials.session_token is None)
+
+    def test_delegate_provider(self):
+        def delegate_get_credentials():
+            return awscrt.auth.AwsCredentials("accesskey", "secretAccessKey", "sessionToken")
+
+        provider = awscrt.auth.AwsCredentialsProvider.new_delegate(delegate_get_credentials)
+        credentials = provider.get_credentials().result(TIMEOUT)
+
+        # Don't use assertEqual(), which could log actual credentials if test fails.
+        self.assertTrue('accesskey' == credentials.access_key_id)
+        self.assertTrue('secretAccessKey' == credentials.secret_access_key)
+        self.assertTrue('sessionToken' == credentials.session_token)
+
+    def test_delegate_provider_exception(self):
+        # delegate that raises exception should result in exception
+        def delegate_get_credentials():
+            raise Exception("purposefully thrown exception")
+
+        provider = awscrt.auth.AwsCredentialsProvider.new_delegate(delegate_get_credentials)
+
+        with self.assertRaises(Exception):
+            credentials_future = provider.get_credentials()
+            credentials = credentials_future.result(TIMEOUT)
+
+    def test_delegate_provider_exception_from_bad_return_type(self):
+        # delegate that returns wrong type should result in exception
+        def delegate_get_credentials():
+            return "purposefully return wrong type"
+
+        provider = awscrt.auth.AwsCredentialsProvider.new_delegate(delegate_get_credentials)
+
+        with self.assertRaises(Exception):
+            credentials_future = provider.get_credentials()
+            credentials = credentials_future.result(TIMEOUT)
+
+
+@unittest.skipUnless(os.environ.get('AWS_TEST_MQTT311_COGNITO_IDENTITY') and os.environ.get('AWS_TEST_MQTT311_COGNITO_ENDPOINT'),
+                     'set env var to run test: AWS_TEST_MQTT311_COGNITO_IDENTITY and AWS_TEST_MQTT311_COGNITO_ENDPOINT')
+class CognitoCredentialsProviderTest(NativeResourceTest):
+
+    def test_unauthenticated(self):
+        identity = os.environ.get('AWS_TEST_MQTT311_COGNITO_IDENTITY')
+        identity_endpoint = os.environ.get('AWS_TEST_MQTT311_COGNITO_ENDPOINT')
+        tls_opts = awscrt.io.TlsContextOptions()
+        tls_context = awscrt.io.ClientTlsContext(tls_opts)
+
+        provider = awscrt.auth.AwsCredentialsProvider.new_cognito(
+            endpoint=identity_endpoint,
+            identity=identity,
+            tls_ctx=tls_context
+        )
+
+        self.assertIsNotNone(provider)
+        credentials = provider.get_credentials().result(TIMEOUT)
+        self.assertIsNotNone(credentials)
+
+    def test_cognito_provider_create_exception_bad_login(self):
+        identity = os.environ.get('AWS_TEST_MQTT311_COGNITO_IDENTITY')
+        identity_endpoint = os.environ.get('AWS_TEST_MQTT311_COGNITO_ENDPOINT')
+        tls_opts = awscrt.io.TlsContextOptions()
+        tls_context = awscrt.io.ClientTlsContext(tls_opts)
+
+        with self.assertRaises(Exception):
+            provider = awscrt.auth.AwsCredentialsProvider.new_cognito(
+                endpoint=identity_endpoint,
+                identity=identity,
+                tls_ctx=tls_context,
+                logins=[('provider1', 5), ('provider2', ['List not string'])]
+            )
+
+            self.assertIsNone(provider)
+
+    def test_maximal_create(self):
+        identity = os.environ.get('AWS_TEST_MQTT311_COGNITO_IDENTITY')
+        identity_endpoint = os.environ.get('AWS_TEST_MQTT311_COGNITO_ENDPOINT')
+        tls_opts = awscrt.io.TlsContextOptions()
+        tls_context = awscrt.io.ClientTlsContext(tls_opts)
+
+        provider = awscrt.auth.AwsCredentialsProvider.new_cognito(
+            endpoint=identity_endpoint,
+            identity=identity,
+            tls_ctx=tls_context,
+            logins=[('provider1', 'token1'), ('provide2', 'token2')],
+            custom_role_arn='not-a-real-arn'
+        )
+
+        self.assertIsNotNone(provider)
+
+
+@unittest.skipUnless(os.environ.get('AWS_TESTING_COGNITO_IDENTITY') and os.environ.get('AWS_TEST_MQTT311_COGNITO_ENDPOINT')
+                     and os.environ.get('AWS_TEST_HTTP_PROXY_HOST') and os.environ.get('AWS_TEST_HTTP_PROXY_PORT'),
+                     'set env var to run test: AWS_TESTING_COGNITO_IDENTITY, AWS_TEST_MQTT311_COGNITO_ENDPOINT, '
+                     'AWS_TEST_HTTP_PROXY_HOST, and AWS_TEST_HTTP_PROXY_PORT')
+class CognitoCredentialsProviderProxyTest(NativeResourceTest):
+
+    def test_unauthenticated(self):
+        identity = os.environ.get('AWS_TESTING_COGNITO_IDENTITY')
+        identity_endpoint = os.environ.get('AWS_TEST_MQTT311_COGNITO_ENDPOINT')
+        tls_opts = awscrt.io.TlsContextOptions()
+        tls_context = awscrt.io.ClientTlsContext(tls_opts)
+        http_proxy_options = awscrt.http.HttpProxyOptions(
+            os.environ.get('AWS_TEST_HTTP_PROXY_HOST'),
+            int(os.environ.get('AWS_TEST_HTTP_PROXY_PORT', '0')))
+
+        provider = awscrt.auth.AwsCredentialsProvider.new_cognito(
+            endpoint=identity_endpoint,
+            identity=identity,
+            tls_ctx=tls_context,
+            http_proxy_options=http_proxy_options
+        )
+
+        self.assertIsNotNone(provider)
+        credentials = provider.get_credentials().result(TIMEOUT)
+        self.assertIsNotNone(credentials)
 
 
 class TestSigningConfig(NativeResourceTest):
@@ -347,3 +545,26 @@ class TestSigner(NativeResourceTest):
                  '20150830T123600Z'),
                 ('Authorization',
                  'AWS4-HMAC-SHA256 Credential=AKIDEXAMPLE/20150830/us-east-1/service/aws4_request, SignedHeaders=host;x-amz-date, Signature=8318018e0b0f223aa2bbf98705b62bb787dc9c0e678f255a891fd03141be5d85')])
+
+    def test_signing_sigv4a_headers(self):
+        credentials_provider = awscrt.auth.AwsCredentialsProvider.new_static(
+            SIGV4TEST_ACCESS_KEY_ID, SIGV4TEST_SECRET_ACCESS_KEY, SIGV4TEST_SESSION_TOKEN)
+
+        signing_config = awscrt.auth.AwsSigningConfig(
+            algorithm=awscrt.auth.AwsSigningAlgorithm.V4_ASYMMETRIC,
+            signature_type=awscrt.auth.AwsSignatureType.HTTP_REQUEST_HEADERS,
+            credentials_provider=credentials_provider,
+            region=SIGV4TEST_REGION,
+            service=SIGV4TEST_SERVICE,
+            date=SIGV4TEST_DATE)
+
+        http_request = awscrt.http.HttpRequest('GET', '/')
+        http_request.headers.add('Host', 'example.amazonaws.com')
+
+        signing_future = awscrt.auth.aws_sign_request(http_request, signing_config)
+        signed_request = signing_future.result(TIMEOUT)
+
+        auth_header_value = signed_request.headers.get('Authorization')
+        self.assertIsNotNone(auth_header_value)
+        self.assertTrue(auth_header_value.startswith(
+            'AWS4-ECDSA-P256-SHA256 Credential=AKIDEXAMPLE/20150830/service/aws4_request, SignedHeaders=host;x-amz-date;x-amz-region-set, Signature='))
